Update deprecated assert(string) calls
In PHP 7.x assert() needs an expression as first argument, not a
string. Some error messages may be less explaining now, because
they won't display the expression itself as they did with string.
They could be improved with providing an Exception object.
Also removed a USE_ASSERTS constant which helped in PHP 3.x.